WIP: drm/msm: Add support for the AQE on a750
The Application QRISC Engine or AQE is a new engine added with a750. It
allows complex work-generating scheduling algorithms to be run on the
GPU. The userspace driver requests the use of the AQE via a special PM4
packet which is forwarded to the AQE, and then the BR/LPAC processor
switches into a special mode where it recieves packets from the AQE
until the AQE is finished. The AQE reads buffers written by shaders and
executes compute dispatches or draw calls in response. There are
currently two "Applications": ray tracing pipelines, and mesh shaders.

From the kernel's perspective, all we need to do is make sure that a
preemption record is allocated, since AQE uses part of it as a scratch
space (e.g. for writing PM4 packets), write the base address of the
firmware to the right register, and handle error interrupts. The SQE
will initialize the AQE in CP_SET_PSEUDO_REG if it has a non-zero
firmware base address, and then userspace can call it without any kernel
intervention.

There are two AQE instances, one used by BR and one used by LPAC. Since
we don't support LPAC, this only adds support for the BR one.

Signed-off-by: Connor Abbott <cwabbott0@gmail.com>