drm/msm: Remove spurious resume in fault handler
The point of enabling stall-on-fault in SMMU is to wait until after
dumping the GPU state in the fault worker to resume the faulty
translation, so when a resume was accidentally added in the page fault
hander in 8cceb773f565 ("drm/msm/adreno: stall translation on fault for
all GPU families") it broke coredumps for GPU page faults. Even worse,
it led to a storm of SMMU context fault IRQs before the fault worker
could be scheduled.

Note that a5xx also calls resume in its pagefault handler, so removing
this should be safe on a5xx.

Fixes: 8cceb773f565 ("drm/msm/adreno: stall translation on fault for all GPU families")
Signed-off-by: Connor Abbott <cwabbott0@gmail.com>