drm/msm: Temporarily disable stall-on-fault after a page fault
When things go wrong, the GPU is capable of quickly generating millions
of faulting translation requests per second. When that happens, in the
stall-on-fault model each access will stall until it wins the race to
update the FSR bit and then the RESUME register is written. This slows
processing page faults to a crawl as the GPU can generate faults much
faster than the CPU can acknowledge them, and the CPU processing the
interrupts is saturated with SMMU interrupts so the rest of the system
locks up. It also means that all available resources in the SMMU are
saturated waiting for the stalled transactions, so that other
transactions such as transactions generated by the GMU (GPU Managment
Unit), which shares a context bank with the GPU, as it responds to power
managments requests from the host, cannot proceed and we get GMU
timeouts. Finally, it prevents power collapse because GX cannot collapse
when there is a transaction pending.

On older platforms with qcom,smmu-v2, it seems that when one transaction
is stalled subsequent faulting transactions are terminated, which avoids
this problem, but the MMU-500 follows the spec here.

To work around the first and third problem, disable stall-on-fault as
soon as we get a page fault until a cooldown period after pagefaults
stop. We also keep it disabled so long as the current devcoredump hasn't
been deleted.

The second problem, of the GMU being stalled and timing out, still
happens occasionally, but the system appears to recover from the error.

Signed-off-by: Connor Abbott <cwabbott0@gmail.com>