drm/msm: Always recover the GPU after a hang
It's possible that the GPU advances past a hang if the GPU is idle,
if the GPU recovers itself. Previously the code would just ignore the
hang, but for some hangs like RBBM GPC errors we need to put the GPU in
an unsafe state by disabling interrupts and the only way to fix it
is to reset the GPU and re-enable interrupts.

Signed-off-by: Connor Abbott <cwabbott0@gmail.com>